tests/wpt/tests/css/css-text/parsing/text-align-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-align with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-align">
<meta name="assert" content="text-align supports only the grammar 'start | end | left | right | center | justify | match-parent | justify-all'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("text-align", "auto");
test_invalid_value("text-align", "start end");
test_invalid_value("text-align", "left right");
test_invalid_value("text-align", "center justify");
test_invalid_value("text-align", "justify-all match-parent");
</script>
</body>
</html>
